merges OpensciencePerformanceCard into StaticPerformanceCard to avoid duplication
