+++ packages/react-components/src/organisms/StaticPerformanceCard.tsx
@@ -1,10 +1,11 @@
 import { css } from '@emotion/react';
-import { Subtitle } from '../atoms';
+import { Paragraph, Subtitle } from '../atoms';
 import { lead } from '../colors';
 import {
   happyFaceIcon,
   informationInverseIcon,
   neutralFaceIcon,
+  PercentageIcon,
   sadFaceIcon,
 } from '../icons';
 import { contentSidePaddingWithNavigation } from '../layout';
@@ -12,14 +13,18 @@ import { rem, smallDesktopScreen } from '../pixels';
 import CaptionCard from './CaptionCard';
 
 const containerStyles = css({
-  display: 'grid',
-  gridTemplateColumns: 'repeat(4, auto)',
-  columnGap: 'clamp(12px, 78px, 78px)',
-  rowGap: rem(16),
+  display: 'flex',
+  justifyContent: 'space-between',
+  flexWrap: 'wrap',
+  gridColumn: '1 / span 2',
+  paddingRight: rem(48),
   [`@media (max-width: calc(${
     smallDesktopScreen.width
-  }px + 220px + ${contentSidePaddingWithNavigation(2)}))`]: {
-    gridTemplateColumns: 'repeat(2, auto)',
+  }px + ${contentSidePaddingWithNavigation(2)}))`]: {
+    display: 'grid',
+    gridTemplateColumns: '1fr 1fr',
+    gap: rem(12),
+    justifyContent: 'stretch',
   },
 });
 
@@ -38,29 +43,73 @@ const dataTextStyles = css({
   color: lead.rgb,
 });
 
-const StaticPerformanceCard: React.FC = () => (
+const captionLegendStyles = css({
+  gridColumn: '1 / span 2',
+  display: 'flex',
+  alignItems: 'flex-start',
+  gap: rem(4),
+});
+
+const iconContainerStyles = css({
+  flexShrink: 0,
+});
+
+const paragraphContainerStyles = css({
+  flex: 1,
+  minWidth: 0,
+});
+
+const StaticPerformanceCard = ({
+  outstanding = 90,
+  adequate = 80,
+  needsImprovement = 80,
+  limitedData = 0,
+  legend,
+}: {
+  outstanding?: number;
+  adequate?: number;
+  needsImprovement?: number;
+  limitedData?: number;
+  legend?: string;
+}) => (
   <CaptionCard>
     <div css={containerStyles}>
       <div css={dataContainerStyles}>
         {happyFaceIcon}
         <Subtitle noMargin>Outstanding:</Subtitle>
-        <span css={dataTextStyles}>&ge; 90%</span>
+        <span css={dataTextStyles}>&ge; {outstanding}%</span>
       </div>
       <div css={dataContainerStyles}>
         {neutralFaceIcon}
         <Subtitle noMargin>Adequate:</Subtitle>
-        <span css={dataTextStyles}>80% - 89%</span>
+        <span css={dataTextStyles}>
+          {adequate}% - {needsImprovement - 1}%
+        </span>
       </div>
       <div css={dataContainerStyles}>
         {sadFaceIcon}
         <Subtitle noMargin>Needs Improvement:</Subtitle>
-        <span css={dataTextStyles}>&lt; 80%</span>
+        <span css={dataTextStyles}>&lt; {needsImprovement}%</span>
       </div>
       <div css={dataContainerStyles}>
         {informationInverseIcon}
         <Subtitle noMargin>Limited data</Subtitle>
       </div>
     </div>
+    {legend && (
+      <div css={captionLegendStyles}>
+        <div css={iconContainerStyles}>
+          {PercentageIcon({ title: 'percentage', color: lead.rgb })}
+        </div>
+        <Paragraph
+          accent="lead"
+          noMargin
+          styles={css(paragraphContainerStyles)}
+        >
+          {legend}
+        </Paragraph>
+      </div>
+    )}
   </CaptionCard>
 );
